+++ go/vt/sqlparser/token.go
@@ -20,11 +20,7 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"strconv"
 	"strings"
-
-	"vitess.io/vitess/go/vt/proto/vtrpc"
-	"vitess.io/vitess/go/vt/vterrors"
 
 	"vitess.io/vitess/go/bytes2"
 	"vitess.io/vitess/go/sqltypes"
@@ -1040,44 +1036,6 @@ func (tkn *Tokenizer) scanMySQLSpecificComment() (int, []byte) {
 	}
 
 	return tkn.Scan()
-}
-
-// ConvertMySQLVersionToCommentVersion converts the MySQL version into comment version format.
-func ConvertMySQLVersionToCommentVersion(version string) (string, error) {
-	var res = make([]int, 3)
-	idx := 0
-	val := ""
-	for _, c := range version {
-		if c <= '9' && c >= '0' {
-			val += string(c)
-		} else if c == '.' {
-			v, err := strconv.Atoi(val)
-			if err != nil {
-				return "", err
-			}
-			val = ""
-			res[idx] = v
-			idx++
-			if idx == 3 {
-				break
-			}
-		} else {
-			break
-		}
-	}
-	if val != "" {
-		v, err := strconv.Atoi(val)
-		if err != nil {
-			return "", err
-		}
-		res[idx] = v
-		idx++
-	}
-	if idx == 0 {
-		return "", vterrors.Errorf(vtrpc.Code_INVALID_ARGUMENT, "MySQL version not correctly setup - %s.", version)
-	}
-
-	return fmt.Sprintf("%01d%02d%02d", res[0], res[1], res[2]), nil
 }
 
 func (tkn *Tokenizer) consumeNext(buffer *bytes2.Buffer) {
